requirements.txt
aiohttp==3.7.4.post0
apipkg==1.5
appdirs==1.4.4
astroid==2.7.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
azure-common==1.1.27
azure-core==1.18.0
azure-identity==1.6.1
azure-mgmt-core==1.3.0
azure-mgmt-resource==20.0.0
azure-mgmt-storage==19.0.0
azure-storage-blob==12.9.0
bandit==1.6.2
beautifulsoup4==4.8.2
black==21.6b0
bleach==3.2.1
Brotli==1.0.9
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coloredlogs==15.0.1
cryptography==3.4.6
cwrap==1.6.1
cycler==0.10.0
dash==2.0.0
dash-bootstrap-components==0.10.7
dash-core-components==2.0.0
dash-dangerously-set-inner-html==0.0.2
dash-daq==0.5.0
dash-flow-example==0.0.5
dash-html-components==2.0.0
dash-pivottable==0.0.2
dash-renderer==1.8.3
dash-table==5.0.0
decorator==5.0.6
defusedxml==0.6.0
Deprecated==1.2.13
deprecation==2.1.0
disjoint-set==0.7.2
ecl==2.9.3
ecl-data-io==0.0.2
ecl2df==0.15.0
entrypoints==0.3
et-xmlfile==1.0.1
execnet==1.7.1
fire==0.4.0
flake8==3.9.2
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.8.0
flask-talisman==0.8.1
fmu-ensemble==1.4.0
fmu-tools==1.8.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.24
h5py==3.2.1
hdf5plugin==2.3.2
humanfriendly==9.1
idna==2.8
importlib-resources==5.4.0
iniconfig==1.1.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jsonschema==4.2.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libecl==2.9.3
lxml==4.5.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mock==4.0.3
more-itertools==8.6.0
msal==1.9.0
msal-extensions==0.3.0
msrest==0.6.21
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
numexpr==2.7.3
numpy==1.19.4
oauthlib==3.1.1
openpyxl==3.0.6
opm==2021.4.0
orjson==3.6.1
packaging==20.4
pandas==1.2.2
pathspec==0.9.0
patsy==0.5.1
pbr==5.5.1
percy==2.0.2
Pillow==8.0.1
pkg_resources==0.0.0
platformdirs==2.3.0
plotly==5.3.1
pluggy==0.13.1
portalocker==1.7.1
py==1.9.0
pyarrow==5.0.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyGithub==1.55
PyJWT==2.0.1
pylint==2.10.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyscal==0.7.7
PySimpleGUI==4.55.1
pystray==0.18.0
pytest==6.0.1
pytest-forked==1.3.0
pytest-mock==3.2.0
pytest-sugar==0.9.4
pytest-xdist==2.1.0
python-dateutil==2.8.1
python-xlib==0.31
pytz==2020.4
PyYAML==5.4.1
regex==2020.11.13
requests==2.21.0
requests-oauthlib==1.3.0
retrying==1.3.3
RoffIO==0.0.2
scipy==1.5.4
segyio==1.9.3
selenium==3.141.0
Shapely==1.7.1
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
statsmodels==0.12.2
stevedore==3.2.2
tables==3.6.1
tenacity==8.0.1
termcolor==1.1.0
toml==0.10.2
tomli==1.2.1
tqdm==4.51.0
treelib==1.6.1
typed-ast==1.4.1
types-bleach==4.1.0
types-Markdown==3.3.3
types-PyYAML==5.4.10
types-requests==2.25.7
typing-extensions==3.10.0.2
urllib3==1.24.3
waitress==1.4.3
watchdog==2.1.6
webencodings==0.5.1
-e git+https://github.com/equinor/webviz-config.git@b3e2b423f76b81a245f90b1b8b5cca7161dff70e#egg=webviz_config
-e git+https://github.com/equinor/webviz-config-equinor.git@796df8515cf3ec9ccfbb6445c7c0f0782e472073#egg=webviz_config_equinor
-e git+https://github.com/equinor/webviz-core-components.git@5a4a534e80fa10eaf60df9ad6d4506975380b97b#egg=webviz_core_components
-e git+ssh://git@github.com/CeetronSolutions/webviz-dev-sync-1.git@66d75a1d0f97c83b23d645cde4c92794d904a8de#egg=webviz_dev_sync
# Editable install with no version control (webviz-local-tests==0.0.0)
-e /home/ruben/git-repos/webviz/webviz-local-tests
-e git+https://github.com/equinor/webviz-subsurface.git@39ad4bc6e64d4f0036ba98faaf27808b97514110#egg=webviz_subsurface
-e git+https://github.com/equinor/webviz-subsurface-components.git@bfaa27b5e46d3decb58240b87d48213ec11baf1f#egg=webviz_subsurface_components
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xtgeo==2.15.2
yarl==1.6.3
zipp==3.6.0
